pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vgMath"
version = "0.1.0"
description = "A math library for Python"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Adrián R. Espinosa" }
]
keywords = ["noise", "procedural", "generation", "perlin", "simplex"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.24.0",
    "numba>=0.58.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "Pillow>=10.0.0",
]
viewer = [
    "Pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/virigir/vgNoise"
Repository = "https://github.com/virigir/vgNoise"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
vgmath = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
